Module piet::conv

Conversions of fundamental numeric and geometric types.

Traits

RoundFrom
This is our own implementation of a "lossy From" trait, representing a conversion that can have precision loss. It is essentially adapted from https://github.com/rust-lang/rfcs/pull/2484.

RoundInto
The companion to RoundFrom. As with From and Into, a blanket implementation is provided; for the most part, implement RoundFrom.