Struct piet::LinearGradient

pub struct LinearGradient { /* fields omitted */ }

A description of a linear gradient in the unit rect, which can be resolved to a fixed gradient.

The start and end points in the gradient are given in UnitPoint coordinates, which are then resolved to image-space coordinates for any given concrete Rect.

When the fixed coordinates are known, use FixedLinearGradient instead.

Implementations

impl LinearGradient

pub fn new(
start: UnitPoint,
end: UnitPoint,
stops: impl GradientStops
) -> LinearGradient

Create a new linear gradient.

The start and end coordinates are UnitPoint coordinates, relative to the geometry of the shape being drawn.

Examples

use piet::{Color, RenderContext, LinearGradient, UnitPoint};
use piet::kurbo::{Circle, Point};

let circle = Circle::new(Point::new(100.0, 100.0), 50.0);
let gradient = LinearGradient::new(
    UnitPoint::TOP,
    UnitPoint::BOTTOM,
    (Color::WHITE, Color::BLACK)
);
render_ctx.fill(circle, &gradient);