Trait piet::IntoBrush

pub trait IntoBrush<P> where
    P: RenderContext + ?Sized, {
    fn make_brush<'a>(
        &'a self, 
        piet: &mut P, 
        bbox: impl FnOnce() -> Rect
    ) -> Cow<'a, P::Brush>;
}

A trait for various types that can be used as brushes. These include backend-independent types such Color and LinearGradient, as well as the types used to represent these on a specific backend.

This is an internal trait that you should not have to implement or think about.

Required methods

fn make_brush<'a>(
&'a self,
piet: &mut P,
bbox: impl FnOnce() -> Rect
) -> Cow<'a, P::Brush>

Loading content...