Module piet::util

utilities shared by various backends Code useful for multiple backends

Functions

compute_blurred_rect
Generate image for a blurred rect, writing it into the provided buffer.

count_until_utf16
returns utf8 text position (code unit offset) at the given utf-16 text position

count_utf16
Counts the number of utf-16 code units in the given string. from xi-editor

size_for_blurred_rect